Move BackupStats to backup.rs
