+++ src/blockdir.rs
@@ -37,11 +37,12 @@ use serde::{Deserialize, Serialize};
 #[allow(unused_imports)]
 use tracing::{debug, error, info, warn};
 
+use crate::backup::BackupStats;
 use crate::blockhash::BlockHash;
 use crate::compress::snappy::{Compressor, Decompressor};
 use crate::kind::Kind;
 use crate::progress::{Bar, Progress};
-use crate::stats::{BackupStats, Sizes};
+use crate::stats::Sizes;
 use crate::transport::local::LocalTransport;
 use crate::transport::{DirEntry, ListDirNames, Transport};
 use crate::*;
