+++ src/lib.rs
@@ -54,9 +54,8 @@ pub mod validate;
 pub use crate::apath::Apath;
 pub use crate::archive::Archive;
 pub use crate::archive::DeleteOptions;
-pub use crate::backup::{backup, BackupOptions};
-pub use crate::band::Band;
-pub use crate::band::BandSelectionPolicy;
+pub use crate::backup::{backup, BackupOptions, BackupStats};
+pub use crate::band::{Band, BandSelectionPolicy};
 pub use crate::bandid::BandId;
 pub use crate::blockdir::BlockDir;
 pub use crate::blockhash::BlockHash;
@@ -74,7 +73,7 @@ pub use crate::misc::bytes_to_human_mb;
 pub use crate::owner::Owner;
 pub use crate::restore::{restore, RestoreOptions};
 pub use crate::show::{show_versions, ShowVersionsOptions};
-pub use crate::stats::{BackupStats, DeleteStats, RestoreStats};
+pub use crate::stats::{DeleteStats, RestoreStats};
 pub use crate::stored_tree::StoredTree;
 pub use crate::transport::{open_transport, Transport};
 pub use crate::tree::{ReadBlocks, ReadTree, TreeSize};
