+++ src/stats.rs
@@ -32,11 +32,15 @@ fn ratio(uncompressed: u64, compressed: u64) -> f64 {
     }
 }
 
-fn write_size<I: Into<u64>>(w: &mut fmt::Formatter<'_>, label: &str, value: I) {
+pub(crate) fn write_size<I: Into<u64>>(w: &mut fmt::Formatter<'_>, label: &str, value: I) {
     writeln!(w, "{:>12} MB   {}", mb_string(value.into()), label).unwrap();
 }
 
-fn write_compressed_size(w: &mut fmt::Formatter<'_>, compressed: u64, uncompressed: u64) {
+pub(crate) fn write_compressed_size(
+    w: &mut fmt::Formatter<'_>,
+    compressed: u64,
+    uncompressed: u64,
+) {
     write_size(w, "uncompressed", uncompressed);
     write_size(
         w,
@@ -45,7 +49,7 @@ fn write_compressed_size(w: &mut fmt::Formatter<'_>, compressed: u64, uncompress
     );
 }
 
-fn write_count<I: Into<usize>>(w: &mut fmt::Formatter<'_>, label: &str, value: I) {
+pub(crate) fn write_count<I: Into<usize>>(w: &mut fmt::Formatter<'_>, label: &str, value: I) {
     writeln!(
         w,
         "{:>12}      {}",
@@ -55,7 +59,11 @@ fn write_count<I: Into<usize>>(w: &mut fmt::Formatter<'_>, label: &str, value: I
     .unwrap();
 }
 
-fn write_duration(w: &mut fmt::Formatter<'_>, label: &str, duration: Duration) -> fmt::Result {
+pub(crate) fn write_duration(
+    w: &mut fmt::Formatter<'_>,
+    label: &str,
+    duration: Duration,
+) -> fmt::Result {
     writeln!(w, "{:>12}      {}", duration_to_hms(duration), label)
 }
 
@@ -112,81 +120,6 @@ impl fmt::Display for RestoreStats {
         write_count(w, "symlinks", self.symlinks);
         write_count(w, "directories", self.directories);
         write_count(w, "unsupported file kind", self.unknown_kind);
-        writeln!(w).unwrap();
-
-        write_count(w, "errors", self.errors);
-        write_duration(w, "elapsed", self.elapsed)?;
-
-        Ok(())
-    }
-}
-
-#[derive(Add, AddAssign, Debug, Default, Eq, PartialEq, Clone)]
-pub struct BackupStats {
-    // TODO: Have separate more-specific stats for backup and restore, and then
-    // each can have a single Display method.
-    // TODO: Include source file bytes, including unmodified files.
-    pub files: usize,
-    pub symlinks: usize,
-    pub directories: usize,
-    pub unknown_kind: usize,
-
-    pub unmodified_files: usize,
-    pub modified_files: usize,
-    pub new_files: usize,
-
-    /// Bytes that matched an existing block.
-    pub deduplicated_bytes: u64,
-    /// Bytes that were stored as new blocks, before compression.
-    pub uncompressed_bytes: u64,
-    pub compressed_bytes: u64,
-
-    pub deduplicated_blocks: usize,
-    pub written_blocks: usize,
-    /// Blocks containing combined small files.
-    pub combined_blocks: usize,
-
-    pub empty_files: usize,
-    pub small_combined_files: usize,
-    pub single_block_files: usize,
-    pub multi_block_files: usize,
-
-    pub errors: usize,
-
-    pub index_builder_stats: IndexWriterStats,
-    pub elapsed: Duration,
-}
-
-impl fmt::Display for BackupStats {
-    fn fmt(&self, w: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write_count(w, "files:", self.files);
-        write_count(w, "  unmodified files", self.unmodified_files);
-        write_count(w, "  modified files", self.modified_files);
-        write_count(w, "  new files", self.new_files);
-        write_count(w, "symlinks", self.symlinks);
-        write_count(w, "directories", self.directories);
-        write_count(w, "unsupported file kind", self.unknown_kind);
-        writeln!(w).unwrap();
-
-        write_count(w, "files stored:", self.new_files + self.modified_files);
-        write_count(w, "  empty files", self.empty_files);
-        write_count(w, "  small combined files", self.small_combined_files);
-        write_count(w, "  single block files", self.single_block_files);
-        write_count(w, "  multi-block files", self.multi_block_files);
-        writeln!(w).unwrap();
-
-        write_count(w, "data blocks deduplicated:", self.deduplicated_blocks);
-        write_size(w, "  saved", self.deduplicated_bytes);
-        writeln!(w).unwrap();
-
-        write_count(w, "new data blocks written:", self.written_blocks);
-        write_count(w, "  blocks of combined files", self.combined_blocks);
-        write_compressed_size(w, self.compressed_bytes, self.uncompressed_bytes);
-        writeln!(w).unwrap();
-
-        let idx = &self.index_builder_stats;
-        write_count(w, "new index hunks", idx.index_hunks);
-        write_compressed_size(w, idx.compressed_index_bytes, idx.uncompressed_index_bytes);
         writeln!(w).unwrap();
 
         write_count(w, "errors", self.errors);
